+++ src/extension_commands.ml
@@ -609,7 +609,7 @@ module Construct = struct
     match selected_result with
     | Some (value, range) -> (
       let* value_inserted = insert_to_document text_editor range value in
-      match Settings.(get server_chainConstructResults_setting) with
+      match Settings.(get server_constructRecursiveCalls_setting) with
       | Some true | None -> (
         match value_inserted with
         | true -> (