+++ src/settings.mli
@@ -46,4 +46,4 @@ val server_duneDiagnostics_setting : bool setting
 
 val server_syntaxDocumentation_setting : bool setting
 
-val server_chainConstructResults_setting : bool setting
+val server_constructRecursiveCalls_setting : bool setting